test: removing user_name duplicate test
You cannot verify via the schema that there is no duplicate user_name
because of a race condition before the user_name is indexed. If another
duplicate user_name happens to be created before the index is created,
the validation will pass, yet there will be a duplicate user_name.

Because of this, I decided to create a wrapper around User to enforce
validation and expose controlled methods to create and manipulate Users.